Editing
Changed various verbs and subjects to match number. Independent clauses not joined by a comma, used a semicolon instead. Joined together dependent clauses with commas.

Criticism for this part of the writing process: I personally do revisions and edits at the same time, so having the revision and edit process separate makes it difficult to document the changes. There doesn't seem to be much to say in this part of the process because a lot of the changes and edits I made already happened in the revision process.